finished exercise. recv and send implemented. Some errorhandling implemented.

+++ Exercises/Exercise1-WrapperFacade/Exercise1-WrapperFacade/INET_Address.hpp
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 #ifdef _WIN32
-#include <winsock.h>
+#include <winsock2.h>
 #else
 //TODO: sockaddr and sockaddr_in
 #endif
@@ -11,9 +11,9 @@
 class INET_Address
 {
 public:
-	INET_Address(uint16_t port, uint64_t addr);
+	INET_Address(uint16_t port, uint32_t addr);
 	uint16_t get_port() const ;
-	uint64_t get_ip_addr() const;
+	uint32_t get_ip_addr() const;
 	const sockaddr *addr() const;
 	size_t size() const;
